добавляет модули для js в проект
