+++ js/constants.js
@@ -1,0 +1,22 @@
+export const PHOTO_COUNT = 25;
+
+export const PHOTO_ID_MIN = 1; //id не должны повторяться
+export const PHOTO_ID_MAX = 25;
+
+export const PHOTO_URL_MIN = 1; //не должны повторяться
+export const PHOTO_URL_MAX = 25;
+
+export const LIKES_MIN = 15;
+export const LIKES_MAX = 200;
+
+export const COMMENT_COUNT_MIN = 0;
+export const COMMENT_COUNT_MAX = 30;
+
+export const COMMENT_ID_MIN = 1; //id не должны повторяться
+export const COMMENT_ID_MAX = 1000;
+
+export const AVATAR_INDEX_MIN = 1;
+export const AVATAR_INDEX_MAX = 6;
+
+export const MESSAGE_COUNT_MIN = 1;
+export const MESSAGE_COUNT_MAX = 2;
