+++ js/data.js
@@ -1,0 +1,25 @@
+const MESSAGES = [
+  'Всё отлично! В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.',
+  'В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают.',
+  'Как можно было поймать такой неудачный момент?!'
+];
+
+//Список имен
+const NAMES = ['Иван', 'Мария', 'Таня', 'Кристоф', 'Виктор', 'Юлия', 'Лилия', 'Валера', 'Юля', 'Андрей', 'Сергей', 'Алексей'];
+
+//Список описаний фотографий
+const DESCRIPTIONS = [
+  'Изображение моря и пляжа',
+  'Изображение девушки с фотоаппаратом',
+  'Изображение блюда с рисом',
+  'Изображение черной машины',
+  'Изображение дрона',
+  'Изображение котика',
+  'Изображение самолета'
+];
+
+export {MESSAGES, NAMES, DESCRIPTIONS};
